+++ tests/arduino_stubs/Udp.h
@@ -3,6 +3,7 @@
 
 #include <stddef.h>
 #include <stdint.h>
+#include "Arduino.h"
 
 class UDP {
   public:
@@ -13,6 +14,8 @@ class UDP {
     virtual int endPacket() = 0;
     virtual size_t write(const uint8_t* data, size_t length) = 0;
     virtual int parsePacket() = 0;
+    virtual IPAddress remoteIP() = 0;
+    virtual uint16_t remotePort() = 0;
     virtual int available() = 0;
     virtual int read(uint8_t* data, size_t length) = 0;
 };